new uniform options; moved cosmetics class file
