+++ src/main/scala/net/psforever/packet/game/objectcreate/CharacterData.scala
@@ -1,8 +1,9 @@
 // Copyright (c) 2017 PSForever
 package net.psforever.packet.game.objectcreate
 
-import net.psforever.objects.avatar.{BattleRank, Cosmetic}
+import net.psforever.objects.avatar.BattleRank
 import net.psforever.packet.{Marshallable, PacketHelpers}
+import net.psforever.types.{Cosmetic, UniformStyle}
 import scodec.codecs._
 import scodec.{Attempt, Codec, Err}
 import shapeless.{::, HNil}
@@ -24,24 +25,6 @@ object ImplantEffects extends Enumeration {
   val NoEffects             = Value(1)
 
   implicit val codec = PacketHelpers.createEnumerationCodec(this, uint4L)
-}
-
-/**
-  * Values for the four different color designs that impact a player's uniform.
-  * Exo-suits get minor graphical updates at the following battle rank levels: seven (1), fourteen (2), and twenty-five (4).
-  * The values 3 and 5 also exist and are visually descriptive to the third upgrade.
-  */
-object UniformStyle extends Enumeration {
-  type Type = Value
-
-  val Normal          = Value(0)
-  val FirstUpgrade    = Value(1)
-  val SecondUpgrade   = Value(2)
-  val SecondUpgradeEx = Value(3)
-  val ThirdUpgrade    = Value(4)
-  val ThirdUpgradeEx  = Value(5)
-
-  implicit val codec = PacketHelpers.createEnumerationCodec(this, uint(3))
 }
 
 /**
@@ -70,9 +53,9 @@ object UniformStyle extends Enumeration {
   *                     cosmetic armor associated with the command rank will be applied automatically
   * @param implant_effects the effects of implants that can be seen on a player's character;
   *                        the number of entries equates to the number of effects applied;
-  *                        the maximu number of effects is three
+  *                        the maximum number of effects is three
   * @param cosmetics optional decorative features that are added to the player's head model by console/chat commands;
-  *                  they become available at battle rank 24, but here they require the third uniform upgrade (rank 25);
+  *                  they become available at battle rank 24;
   *                  these flags do not exist if they are not applicable
   * @param is_backpack this player character should be depicted as a corpse;
   *                    corpses are either coffins (defunct), backpacks (normal), or a pastry (festive);
@@ -83,7 +66,7 @@ object UniformStyle extends Enumeration {
 final case class CharacterData(
     health: Int,
     armor: Int,
-    uniform_upgrade: UniformStyle.Value,
+    uniform_upgrade: UniformStyle,
     unk: Int,
     command_rank: Int,
     implant_effects: List[ImplantEffects.Value],
@@ -124,7 +107,7 @@ object CharacterData extends Marshallable[CharacterData] {
   def apply(
       health: Int,
       armor: Int,
-      uniform: UniformStyle.Value,
+      uniform: UniformStyle,
       cr: Int,
       implant_effects: List[ImplantEffects.Value],
       cosmetics: Option[Set[Cosmetic]]
